+++ test/fixtures/sidekick-config.js
