+++ test/mocks/chrome.js
